learning/requirements.txt
pytorch-lightning
tensorboard
torchvision
torch-summary
